requirements.txt
boto3==1.17.72
botocore==1.20.72
cffi==1.14.5
cryptography==3.4.7
jmespath==0.10.0
pycparser==2.20
python-dateutil==2.8.1
PyYAML==5.4.1
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.4
